refactor(brett): split oversized files to pass S1 line-length gate
Extract asyncHandler to server/helpers.ts, share-link routes to
server/share-routes.ts (index.ts: 628→593 lines). Compact
mountShareButton call in board-boot.ts and trim stale T3 header
comment (610→603 lines). Remove redundant comment block before
connectWS and extra blank line in ws-client.ts (604→599 lines).
S1 gate: 600 lines max; no behavior change.
