+++ brett/src/server/helpers.ts
@@ -1,0 +1,3 @@
+export function asyncHandler(fn: any) {
+  return (req: any, res: any, next: any) => Promise.resolve(fn(req, res, next)).catch(next);
+}
